requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
fastapi==0.108.0
Flask==3.0.0
fonttools==4.47.0
git-lfs==1.6
h11==0.14.0
idna==3.6
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.2
regex==2023.12.25
scikit-learn==1.3.2
scikit-learn-extra==0.3.0
scikit-surprise==1.1.3
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.32.0.post1
threadpoolctl==3.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
uvicorn==0.25.0
wcwidth==0.2.12
Werkzeug==3.0.1
